+++ zenml/core/steps/base_step.py
@@ -15,7 +15,7 @@
 import inspect
 from typing import Dict
 
-from zenml.core.backends.base_backend import BaseBackend
+from zenml.core.backends import BaseBackend
 from zenml.core.standards.standard_keys import StepKeys
 from zenml.utils.enums import StepTypes
 from zenml.utils.print_utils import to_pretty_string, PrintStyles
