+++ zenml/core/steps/base_step_test.py
@@ -12,8 +12,8 @@
 #  or implied. See the License for the specific language governing
 #  permissions and limitations under the License.
 
-from zenml.core.steps.base_step import BaseStep
-from zenml.core.backends.base_backend import BaseBackend
+from zenml.core.backends import BaseBackend
+from zenml.core.steps import BaseStep
 
 
 def test_to_from_config(equal_steps):
