+++ zenml/core/steps/data/bq_data_step.py
@@ -18,7 +18,7 @@ from typing import Text, Any, Dict
 import apache_beam as beam
 from apache_beam.io.gcp import bigquery as beam_bigquery
 
-from zenml.core.steps.data.base_data_step import BaseDataStep
+from zenml.core.steps.data import BaseDataStep
 
 
 @beam.ptransform_fn
