+++ zenml/core/steps/deployer/cortex_deployer.py
@@ -17,8 +17,8 @@ import os
 from typing import Dict, Text, List
 
 from zenml.core.components.pusher import cortex_executor
-from zenml.core.repo.repo import Repository
-from zenml.core.steps.deployer.base_deployer import BaseDeployerStep
+from zenml.core.repo import Repository
+from zenml.core.steps.deployer import BaseDeployerStep
 from zenml.utils.logger import get_logger
 from zenml.utils.source_utils import get_path_from_source, \
     get_class_path_from_source
