+++ zenml/core/steps/evaluator/base_evaluator.py
@@ -13,7 +13,7 @@
 #  permissions and limitations under the License.
 
 
-from zenml.core.steps.base_step import BaseStep
+from zenml.core.steps import BaseStep
 from zenml.utils.enums import StepTypes
