+++ zenml/core/steps/evaluator/tfma_evaluator.py
@@ -19,7 +19,6 @@ import tensorflow_model_analysis as tfma
 from google.protobuf.wrappers_pb2 import BoolValue
 
 from zenml.core.steps.evaluator.base_evaluator import BaseEvaluatorStep
-from zenml.utils.enums import StepTypes
 
 
 def to_camel_case(s):
