+++ zenml/core/steps/split/__init__.py
@@ -1,0 +1,21 @@
+#  Copyright (c) maiot GmbH 2020. All Rights Reserved.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at:
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+#  or implied. See the License for the specific language governing
+#  permissions and limitations under the License.
+
+from zenml.core.steps.split.base_split_step import BaseSplit
+from zenml.core.steps.split.categorical_domain_split_step import \
+    CategoricalDomainSplit
+from zenml.core.steps.split.categorical_ratio_split_step import \
+    CategoricalRatioSplit
+from zenml.core.steps.split.no_split_step import NoSplit
+from zenml.core.steps.split.random_split import RandomSplit
