+++ zenml/core/steps/split/categorical_domain_split_step.py
@@ -16,7 +16,7 @@
 from typing import Text, Dict, List, Any, Union
 
 from zenml.core.steps.split import constants
-from zenml.core.steps.split.base_split_step import BaseSplit
+from zenml.core.steps.split import BaseSplit
 from zenml.core.steps.split.utils import get_categorical_value
 
 CategoricalValue = Union[Text, int]
