+++ zenml/core/steps/split/categorical_ratio_split_step.py
@@ -16,7 +16,7 @@
 from typing import Text, List, Dict, Union
 
 from zenml.core.steps.split import constants
-from zenml.core.steps.split.base_split_step import BaseSplit
+from zenml.core.steps.split import BaseSplit
 from zenml.core.steps.split.categorical_domain_split_step import \
     CategoricalPartitionFn
 from zenml.core.steps.split.utils import partition_cat_list
