+++ zenml/core/steps/split/no_split_step.py
@@ -16,7 +16,7 @@
 from typing import Text, List, Any
 
 from zenml.core.steps.split import constants
-from zenml.core.steps.split.base_split_step import BaseSplit
+from zenml.core.steps.split import BaseSplit
 
 
 def NoSplitPartitionFn(element: Any,
