+++ zenml/core/steps/split/random_split.py
@@ -19,7 +19,7 @@ from typing import Text, List, Dict, Any
 import numpy as np
 
 from zenml.core.steps.split import constants
-from zenml.core.steps.split.base_split_step import BaseSplit
+from zenml.core.steps.split import BaseSplit
 
 
 def lint_split_map(split_map: Dict[Text, float]):
