+++ zenml/core/steps/tokenizer/hf_tokenizer.py
@@ -16,8 +16,8 @@ from typing import Dict, Text, List, Any
 
 import tensorflow as tf
 
-from zenml.core.steps.tokenizer.base_tokenizer import BaseTokenizer
 from zenml.core.steps.split.utils import get_categorical_value
+from zenml.core.steps.tokenizer import BaseTokenizer
 from zenml.core.steps.tokenizer.utils import tokenizer_map
 from zenml.utils import path_utils
