+++ zenml/core/steps/trainer/__init__.py
@@ -1,0 +1,23 @@
+#  Copyright (c) maiot GmbH 2020. All Rights Reserved.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at:
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+#  or implied. See the License for the specific language governing
+#  permissions and limitations under the License.
+
+from zenml.core.steps.trainer.base_trainer import BaseTrainerStep
+from zenml.core.steps.trainer.pytorch_trainers.torch_base_trainer import \
+    TorchBaseTrainerStep
+from zenml.core.steps.trainer.pytorch_trainers.torch_ff_trainer import \
+    FeedForwardTrainer as TorchFeedForwardTrainer
+from zenml.core.steps.trainer.tensorflow_trainers.tf_base_trainer import \
+    TFBaseTrainerStep
+from zenml.core.steps.trainer.tensorflow_trainers.tf_ff_trainer import \
+    FeedForwardTrainer as TFFeedForwardTrainer
